Updated GE Link for new tile layout and added fix for setLevel commands when bulb off.
My take on how the multi-Attribute tile would work for the GE Link. Since most of the relevant info is in the multi-section, I added the current preference selections for easy reference. I kept the old horizontal slider in case users want finer control with the number displayed during adjustment. I also added a "fix" to the setLevel section, so the bulb will always transition smoothly between states (see the latest post in the GE Link thread on the community forum for a more detailed explanation).